Moving repository to a new package
