+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/viewsstats/ViewsStatsViewModel.kt
@@ -10,12 +10,11 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import org.wordpress.android.R
-import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.store.AccountStore
 import org.wordpress.android.ui.mysite.SelectedSiteRepository
-import org.wordpress.android.ui.newstats.todaysstat.StatsRepository
-import org.wordpress.android.ui.newstats.todaysstat.WeeklyAggregates
-import org.wordpress.android.ui.newstats.todaysstat.WeeklyStatsWithDailyDataResult
+import org.wordpress.android.ui.newstats.repository.StatsRepository
+import org.wordpress.android.ui.newstats.repository.WeeklyAggregates
+import org.wordpress.android.ui.newstats.repository.WeeklyStatsWithDailyDataResult
 import org.wordpress.android.viewmodel.ResourceProvider
 import java.text.SimpleDateFormat
 import java.util.Locale
